Begin getting much of the code to start compiling normally
Fixed CMake so that the code will now start to compile the way it
should. I need to fix a few things with the cmake build settings for
C files but that will be next commit. There is a lot of errors still
in the C++ code which needs to be fixed and will take a lot of time
to figure out what is wrong or missing but so far everything should
compile with what is now in the repo. Hopefully the external deps
of this library aren't going to make things be a problem in the future.